update: 更新文件 multicast_set_loopback_context.cpp

+++ frameworks/js/napi/socket/async_context/src/multicast_set_loopback_context.cpp
@@ -84,7 +84,7 @@ int MulticastSetLoopbackContext::GetSocketFd() const
     if (sharedManager_ == nullptr) {
         return -1;
     }
-    return sharedManager_->GetData() ? static_cast<int>(reinterpret_cast<intptr_t>(sharedManager_->GetData())) : -1;
+    return sharedManager_->GetData() ? static_cast<int>(reinterpret_cast<uint64_t>(sharedManager_->GetData())) : -1;
 }
 
 int32_t MulticastSetLoopbackContext::GetErrorCode() const
